Changed support OS and Updated codes/mail addresses

+++ .github/workflows/build_helper.sh
@@ -704,7 +704,7 @@ else
 	fi
 	if [ -z "${CI_DEVELOPER_EMAIL}" ]; then
 		PRNWARN "DEVELOPER_EMAIL is not set in the options or environment variables. There is no choice but to set the default value."
-		CI_DEVELOPER_EMAIL="antpickax-support@mail.yahoo.co.jp"
+		CI_DEVELOPER_EMAIL="ml-antpickax-support@lycorp.co.jp"
 	fi
 fi
 
@@ -1277,7 +1277,7 @@ if [ "${RUN_CPPCHECK}" -eq 1 ]; then
 		fi
 
 	elif [ "${IS_OS_ROCKY}" -eq 1 ]; then
-		if echo "${CI_OSTYPE}" | sed -e 's#:##g' | grep -q -i 'rockylinux8'; then
+		if echo "${CI_OSTYPE}" | sed -e 's#:##g' | grep -q -i 'rockylinux[:]*8'; then
 			#
 			# Rocky 8
 			#
@@ -1293,11 +1293,23 @@ if [ "${RUN_CPPCHECK}" -eq 1 ]; then
 				PRNERR "Failed to enable powertools"
 				exit 1
 			fi
-		else
+		elif echo "${CI_OSTYPE}" | sed -e 's#:##g' | grep -q -i 'rockylinux[:]*9'; then
 			#
 			# Rocky 9 or later
 			#
 			if ({ RUNCMD "${INSTALLER_BIN}" "${INSTALL_CMD}" "${INSTALL_CMD_ARG}" "${INSTALL_AUTO_ARG}" https://dl.fedoraproject.org/pub/epel/epel-release-latest-9.noarch.rpm || echo > "${PIPEFAILURE_FILE}"; } | sed -e 's/^/    /g') && rm "${PIPEFAILURE_FILE}" >/dev/null 2>&1; then
+				PRNERR "Failed to install epel repository"
+				exit 1
+			fi
+			if ({ RUNCMD "${INSTALLER_BIN}" config-manager --enable epel || echo > "${PIPEFAILURE_FILE}"; } | sed -e 's/^/    /g') && rm "${PIPEFAILURE_FILE}" >/dev/null 2>&1; then
+				PRNERR "Failed to enable epel repository"
+				exit 1
+			fi
+		else
+			#
+			# Rocky 10 or later
+			#
+			if ({ RUNCMD "${INSTALLER_BIN}" "${INSTALL_CMD}" "${INSTALL_CMD_ARG}" "${INSTALL_AUTO_ARG}" https://dl.fedoraproject.org/pub/epel/epel-release-latest-10.noarch.rpm || echo > "${PIPEFAILURE_FILE}"; } | sed -e 's/^/    /g') && rm "${PIPEFAILURE_FILE}" >/dev/null 2>&1; then
 				PRNERR "Failed to install epel repository"
 				exit 1
 			fi
